+++ tests/01-register.spec.js
